Refactor: Webclient exception handling, provider
Previous implementation of exception handling for WebClient
did not provide specific information about external API fetching
error. In this commit providing a more specific information about an
error (404, 4xx, 5xx) of fetching Github user data is implemented.

Calculation logic is improved. In the previous implementation case of a
division by zero was not served. In this commit, in a case of the
division by zero, the result of the division is set to zero.

Name of the WebClientBuilder class was renamed to WebClientProvider
to avoid confusion with design patterns.

Minor changes to configuration files are introduced.